pulsar-functions/vectorize-function/vectorize/requirements.txt
python-dotenv
pulsar-client==2.10
